feat!: Add Dashboards v3 client and split Monitor
BREAKING CHANGE:
Splits the SdMonitorClient into a DashboardClient's,
both from version v2 and v3. By default now SdMonitorClient
inherits from DashboardClientV3 and therefore implements
all the interface, but the following methods have changed
the signature:
- add_dashboard_panel
- remove_dashboard_panel

This is required because the Dashboards V3 do not implement
the same kind of panels as V2 implemented.
