+++ Pipfile
@@ -17,6 +17,7 @@ coverage = "*"
 requests = "*"
 pyaml = "*"
 requests-toolbelt = "*"
+tatsu = "*"
 
 [requires]
 python_version = "3.8"
