+++ sdcclient/monitor/__init__.py
@@ -1,2 +1,4 @@
+from ._dashboards_v3 import DashboardsClientV3
+from ._dashboards_v2 import DashboardsClientV2
 from ._events_v1 import EventsClientV1
 from ._events_v2 import EventsClientV2
