+++ sdcclient/monitor/dashboard_converters/__init__.py
@@ -1,0 +1,2 @@
+from ._dashboard_versions import convert_dashboard_between_versions
+from ._dashboard_scope import convert_scope_string_to_expression
